+++ hokusai/cli/registry.py
@@ -15,7 +15,7 @@ def registry(context_settings=CONTEXT_SETTINGS):
 @click.option('--tag', type=click.STRING, help='The remote tag to push to (default: the value of `git rev-parse HEAD`)')
 @click.option('--local-tag', type=click.STRING, default='latest', help='The local tag to push (default: latest)')
 @click.option('--build/--no-build', default=True, help='Force a build of the :latest image before pushing (default: true)')
-@click.option('--filename', type=click.STRING, help='Use the docker-compose Yaml file in the ./hokusai directory (default: build.yml)')
+@click.option('-f', '--filename', type=click.STRING, help='Use the given docker-compose Yaml file when building (default: ./hokusai/build.yml)')
 @click.option('--force', type=click.BOOL, is_flag=True, help='Push even if working directory is not clean')
 @click.option('--overwrite', type=click.BOOL, is_flag=True, help='Push even if the tag already exists')
 @click.option('--skip-latest', type=click.BOOL, is_flag=True, help="Don't update the 'latest' tag" )