+++ hokusai/cli/staging.py
@@ -15,18 +15,18 @@ def staging(context_settings=CONTEXT_SETTINGS):
   pass
 
 @staging.command(context_settings=CONTEXT_SETTINGS)
-@click.option('--filename', type=click.STRING, help='Use the Kubernetes Yaml file in the ./hokusai directory (default staging.yml)')
+@click.option('-f', '--filename', type=click.STRING, help='Use the given Kubernetes Yaml file (default ./hokusai/staging.yml)')
 @click.option('-v', '--verbose', type=click.BOOL, is_flag=True, help='Verbose output')
-def create(verbose):
+def create(filename, verbose):
   """Create the Kubernetes resources defined in ./hokusai/staging.yml"""
-  set_verbosity(filename, verbose)
+  set_verbosity(verbose)
   hokusai.k8s_create(KUBE_CONTEXT, filename=filename)
 
 
 @staging.command(context_settings=CONTEXT_SETTINGS)
-@click.option('--filename', type=click.STRING, help='Use the Kubernetes Yaml file in the ./hokusai directory (default staging.yml)')
+@click.option('-f', '--filename', type=click.STRING, help='Use the given Kubernetes Yaml file (default ./hokusai/staging.yml)')
 @click.option('-v', '--verbose', type=click.BOOL, is_flag=True, help='Verbose output')
-def delete(verbose):
+def delete(filename, verbose):
   """Delete the Kubernetes resources defined in ./hokusai/staging.yml"""
   set_verbosity(verbose)
   hokusai.k8s_delete(KUBE_CONTEXT, filename=filename)
@@ -36,16 +36,15 @@ def delete(verbose):
 @click.option('--check-branch', type=click.STRING, default="master", help='Check branch before updating (default: master)')
 @click.option('--check-remote', type=click.STRING, help='Check remotes before updating (otherwise check all remotes)')
 @click.option('--skip-checks', type=click.BOOL, is_flag=True, help='Skip all checks and update configuration recklessly')
-@click.option('--filename', type=click.STRING, help='Use the Kubernetes Yaml file in the ./hokusai directory (default staging.yml)')
-@click.option('--tag', type=click.STRING, help='Also update the given deployment tag')
+@click.option('-f', '--filename', type=click.STRING, help='Use the given Kubernetes Yaml file (default ./hokusai/staging.yml)')
 @click.option('--dry-run', type=click.BOOL, is_flag=True, help='Perform a dry run of the configuration update')
 @click.option('-v', '--verbose', type=click.BOOL, is_flag=True, help='Verbose output')
-def update(check_branch, check_remote, skip_checks, filename, tag, dry_run, verbose):
+def update(check_branch, check_remote, skip_checks, filename, dry_run, verbose):
   """Update the Kubernetes resources defined in ./hokusai/staging.yml"""
   set_verbosity(verbose)
   hokusai.k8s_update(KUBE_CONTEXT, check_branch=check_branch,
                       check_remote=check_remote, skip_checks=skip_checks,
-                      filename=filename, tag=tag, dry_run=dry_run)
+                      filename=filename, dry_run=dry_run)
 
 
 @staging.command(context_settings=CONTEXT_SETTINGS)
@@ -53,7 +52,7 @@ def update(check_branch, check_remote, skip_checks, filename, tag, dry_run, verb
 @click.option('--pods/--no-pods', default=True, help='Print pods (default: true)')
 @click.option('--describe', type=click.BOOL, is_flag=True, help="Print 'kubectl describe' output for resources and pods")
 @click.option('--top', type=click.BOOL, is_flag=True, help='Print top pods')
-@click.option('--filename', type=click.STRING, help='Use the Kubernetes Yaml file in the ./hokusai directory (default staging.yml)')
+@click.option('-f', '--filename', type=click.STRING, help='Use the given Kubernetes Yaml file (default ./hokusai/staging.yml)')
 @click.option('-v', '--verbose', type=click.BOOL, is_flag=True, help='Verbose output')
 def status(resources, pods, describe, top, filename, verbose):
   """Print Kubernetes resources in the staging context"""
@@ -92,13 +91,15 @@ def logs(timestamps, follow, tail, previous, label, verbose):
 @click.option('--constraint', type=click.STRING, multiple=True, help='Constrain migration and deploy hooks to run on nodes matching labels in the form of "key=value"')
 @click.option('--git-remote', type=click.STRING, help='Push deployment tags to git remote')
 @click.option('-t', '--timeout', type=click.INT, default=600, help="Timeout deployment rollout after N seconds (default 600)")
+@click.option('--update-config', type=click.BOOL, is_flag=True, help='Also update Kubernetes config')
+@click.option('-f', '--filename', type=click.STRING, help='If updating config, use the Kubernetes Yaml file in the ./hokusai directory (default staging.yml)')
 @click.option('-v', '--verbose', type=click.BOOL, is_flag=True, help='Verbose output')
-def deploy(tag, migration, constraint, git_remote, timeout, verbose):
+def deploy(tag, migration, constraint, git_remote, timeout, update_config, filename, verbose):
   """Update the project's deployment(s) to reference
   the given image tag and update the tag staging
   to reference the same image"""
   set_verbosity(verbose)
-  hokusai.update(KUBE_CONTEXT, tag, migration, constraint, git_remote, timeout)
+  hokusai.update(KUBE_CONTEXT, tag, migration, constraint, git_remote, timeout, update_config=update_config, filename=filename)
 
 
 @staging.command(context_settings=CONTEXT_SETTINGS)