+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+## v0.5.6
+
+[Documentation](https://github.com/artsy/hokusai/blob/v0.5.6/README.md)
+
+### Notable changes since v0.5.5
+* c6f8338 Add a short option -f for --filename ( -fi in dev logs command )
+* 07bea52 Add --update-config and --filename options to staging/production deploy
+* 1896d71 Fix ECR integration tests for boto update
+* f2fa240 Update yaml-file-name to filename and refactor commands / fix kubernetes integration tests
+* 1d93ee6 Add a --service-name option to test command to get the return value from a different docker-compose service
+* 017f03c Add a global config file `~/.hokusai/config.yml` to run hokusai configure with no arguments but keep backwards compatibility
+* 1b33220 Add a --dry-run flag to staging/production update and pass through to kubectl apply
+* fdb2e6b Install the aws cli and iam authenticator in the docker image
+* bb74eda Loosen dependencies
+* 88220af Create and delete env configmap with stack
+* 05c4bcf Add the option to override Yaml filename when building, test, dev, staging production and pushing images
+* b6f7725 Add registry pull command
+
+
 ## v0.5.5
 
 [Documentation](https://github.com/artsy/hokusai/blob/v0.5.5/README.md)